REFACTOR-#4696: Merge dask and ray virtual partition code.
